Spelling and grammar fixes of overview.md (#16)
Corrected miscellaneous errors. Altered some sentences.

For future: standardize spelling of "centralize" with a *z*, and "organization" with a *z*. This is the American spelling. (Instead of using *s* sporadically, which is British, and lacks consistency.)

In the photo under *Issues with centralized exchanges*, should it say "Exchange A" and "Exchange B", instead of both showing " Exchange A"? This highlights users' exposure to multiple exchange risks. 
Also in same photo, changing "transfer" to "deposit" may be more clear (and highlights risk).